requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
langchain==0.3.23
langchain-openai==0.3.12
openai==1.74.0
pydantic==2.11.3
python-dotenv==1.1.0
requests==2.31.0
apscheduler==3.10.4
jinja2==3.1.3
gunicorn==23.0.0
email-validator==2.1.0
sqlalchemy==2.0.28
werkzeug==2.3.7